backend/requirements.txt
Django
djangorestframework
pymongo
numpy
pandas
scikit-learn
tensorflow